+++ src/contactsHelpers.ts
@@ -1,35 +1,11 @@
 import { LiveStore, NamedNode, sym } from "rdflib"
 import { ns, utils } from "solid-ui"
-import ContactsModuleRdfLib, { NewContact } from "@solid-data-modules/contacts-rdflib"
+import ContactsModuleRdfLib from "@solid-data-modules/contacts-rdflib"
 import { DataBrowserContext } from "pane-registry";
 import { createAddressBookListDiv, createAddressBookUriSelectorDiv } from "./contactsCards";
 import './styles/contactCards.css'
 import { authn } from "solid-logic";
-export interface SelectedAddressBookUris {
-  addressBookUri: string,
-  groupUris: string []
-}
-export type GroupData = {
-  name: string,
-  uri: string
-}
-interface AddressBookDetails {
-  name: string,
-  groups: GroupData[]
-}
-
-interface AddressBooksData {
-  public: Map<string, AddressBookDetails>,
-  private: Map<string, AddressBookDetails>,
-  contacts: Map<string,string>
-}
-
-export interface ContactData {
-    name: string,
-    email: string[],
-    phoneNumber: string[],
-    webID: string
-}
+import { AddressBooksData, ContactData, SelectedAddressBookUris } from "./contactsTypes";
 
 async function addContactToAddressBook(
   context: DataBrowserContext,
@@ -198,8 +174,6 @@ async function createContactInAddressBook(
 }
 
 export {
-  AddressBooksData,
-  AddressBookDetails,
   getAddressBooksData,
   getContactData,
   addContactToAddressBook,
