+++ src/contactsTypes.ts
@@ -1,0 +1,25 @@
+export interface SelectedAddressBookUris {
+  addressBookUri: string,
+  groupUris: string []
+}
+export type GroupData = {
+  name: string,
+  uri: string
+}
+export interface AddressBookDetails {
+  name: string,
+  groups: GroupData[]
+}
+
+export interface AddressBooksData {
+  public: Map<string, AddressBookDetails>,
+  private: Map<string, AddressBookDetails>,
+  contacts: Map<string,string>
+}
+
+export interface ContactData {
+    name: string,
+    email: string[],
+    phoneNumber: string[],
+    webID: string
+}
